@@ -2,13 +2,12 @@ source 'https://rubygems.org'
 
 
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '4.2.10'
+gem 'rails', '=5.2.3'
+gem 'bootsnap', require: false
 # Use SCSS for stylesheets
 gem 'sass-rails', '~> 5.0'
 # Use Uglifier as compressor for JavaScript assets
 gem 'uglifier', '>= 1.3.0'
-# Use CoffeeScript for .coffee assets and views
-gem 'coffee-rails', '~> 4.1.0'
 # See https://github.com/rails/execjs#readme for more supported runtimes
 # gem 'therubyracer', platforms: :ruby
 # Use jquery as the JavaScript library
@@ -42,7 +41,7 @@ gem 'rails_db'
 if File.exist?("../rock_scissors_paper")
   gem "rock_scissors_paper", :path => "../rock_scissors_paper"
 else
-  gem "rock_scissors_paper", :git => 'https://github.com/Karoid/rock_scissors_paper'
+  gem "rock_scissors_paper", :git => 'https://github.com/Karoid/rock_scissors_paper', :branch => 'rails-5-upgrade'
 end
 
 if File.exist?("../check_attendence")
@@ -88,7 +87,7 @@ end
 
 group :development do
   # Use sqlite3 as the database for Active Record
-  gem 'sqlite3'
+  gem 'sqlite3', '< 1.4'
   # Access an IRB console on exception pages or by using <%= console %> in views
   gem 'web-console', '~> 2.0'
   #Supporting gem for Rails Panel (Google Chrome extension for Rails development).
