@@ -1,5 +1,5 @@
 class Member < ActiveRecord::Base
-  has_many :articles, dependent: :destroy
+  has_many :articles
   has_many :statistics, dependent: :destroy
   has_many :comments, dependent: :destroy
   has_many :kakao_chat_logins, dependent: :destroy
